+++ src/client/cypress/e2e/detailPage/labeling.cy.js
@@ -189,9 +189,7 @@ describe("Test labeling tool", () => {
     isDisabled("locationYLV03", true);
   });
 
-  // TODO: https://github.com/swisstopo/swissgeol-boreholes-suite/issues/1546
-  // api call extract_data in drawBox times out
-  it.skip("can extract data from rotated and zoomed next page", () => {
+  it("can extract data from rotated and zoomed next page", () => {
     newEditableBorehole().as("borehole_id");
     cy.get('[data-cy="labeling-toggle-button"]').click();
     cy.get('[data-cy="labeling-file-dropzone"]').should("exist");
@@ -227,33 +225,25 @@ describe("Test labeling tool", () => {
       expect(view.getRotation()).to.equal(Math.PI / 2);
     });
     cy.wait(1000);
-
     cy.get('[data-cy="labeling-panel"] [data-cy="zoom-in-button"]').click();
     cy.wait(1000);
-    cy.get('[data-cy="labeling-panel"]')
-      .trigger("pointerdown", { x: 600, y: 60 })
-      .trigger("pointerup", { x: 600, y: 200 });
-    cy.wait(1000);
-
-    drawBox(600, 60, 800, 170);
+    drawBox(400, 60, 600, 220);
     evaluateSelect("originalReferenceSystem", "20104002");
-    evaluateCoordinate("locationX", "2'646'466");
-    hasError("locationX", false);
-    isDisabled("locationX", true);
-    evaluateCoordinate("locationY", "1'249'931");
-    hasError("locationY", false);
-    isDisabled("locationY", true);
-    evaluateCoordinate("locationXLV03", "646'465");
+    evaluateCoordinate("locationXLV03", "646'465.97");
     hasError("locationXLV03", false);
     isDisabled("locationXLV03", false);
-    evaluateCoordinate("locationYLV03", "249'931");
+    evaluateCoordinate("locationYLV03", "249'931.66");
     hasError("locationYLV03", false);
     isDisabled("locationYLV03", false);
+    evaluateCoordinate("locationX", "2'646'466.70");
+    hasError("locationX", false);
+    isDisabled("locationX", true);
+    evaluateCoordinate("locationY", "1'249'931.82");
+    hasError("locationY", false);
+    isDisabled("locationY", true);
   });
 
-  // TODO: https://github.com/swisstopo/swissgeol-boreholes-suite/issues/1546
-  // api call extract_data in drawBox times out
-  it.skip("shows alert if no coordinates are extracted", () => {
+  it("shows alert if no coordinates are extracted", () => {
     newEditableBorehole().as("borehole_id");
     cy.get('[data-cy="labeling-toggle-button"]').click();
     cy.get('[data-cy="labeling-file-dropzone"]').should("exist");
@@ -273,6 +263,7 @@ describe("Test labeling tool", () => {
     cy.get('[data-cy="labeling-page-next"]').should("not.be.disabled");
 
     cy.get('[data-cy="labeling-page-next"]').click();
+    cy.get('[data-cy="labeling-page-next"]').click();
     waitForLabelingImageLoaded();
     cy.get('[data-cy="labeling-page-count"]').contains("3 / 3");
     cy.get('[data-cy="labeling-page-previous"]').should("not.be.disabled");
@@ -280,14 +271,9 @@ describe("Test labeling tool", () => {
     cy.wait(1000);
 
     cy.get('[data-cy="coordinate-segment"] [data-cy="labeling-button"]').click();
-    drawBox(50, 60, 400, 110);
-    evaluateSelect("originalReferenceSystem", "20104002");
-    evaluateCoordinate("locationX", "2'646'466");
-    evaluateCoordinate("locationY", "1'249'931");
-    evaluateCoordinate("locationXLV03", "646'465");
-    evaluateCoordinate("locationYLV03", "249'931");
 
-    cy.get('[data-cy="labeling-alert"]').contains("Coordinates not found.");
+    drawBox(180, 70, 400, 110);
+    cy.get('[data-cy="labeling-alert"]').contains("No coordinates found");
 
     // Drawing is active immediately when opening the panel with the labeling-button
     cy.get('[data-cy="labeling-toggle-button"]').click();
@@ -295,13 +281,5 @@ describe("Test labeling tool", () => {
 
     cy.get('[data-cy="coordinate-segment"] [data-cy="labeling-button"]').click();
     cy.get('[data-cy="labeling-file-dropzone"]').should("exist");
-    cy.contains("labeling_attachment.pdf").click();
-    waitForLabelingImageLoaded();
-    drawBox(400, 60, 600, 170);
-    evaluateSelect("originalReferenceSystem", "20104001");
-    evaluateCoordinate("locationX", "2'646'359.7");
-    evaluateCoordinate("locationY", "1'249'017.82");
-    evaluateCoordinate("locationXLV03", "646'358.97");
-    evaluateCoordinate("locationYLV03", "249'017.66");
   });
 });