+++ src/client/src/pages/detail/detailPage.tsx
@@ -27,6 +27,7 @@ export const DetailPage: FC = () => {
   const [updatedBy, setUpdatedBy] = useState<User | null>(null);
   const legacyBorehole: Borehole = useSelector((state: ReduxRootState) => state.core_borehole);
   const user = useSelector((state: ReduxRootState) => state.core_user);
+  const workflowStatus = useSelector((state: ReduxRootState) => state.core_workflow);
   const location = useLocation();
   const { panelPosition, panelOpen, togglePanel } = useLabelingContext();
   const dispatch = useDispatch();
@@ -39,7 +40,7 @@ export const DetailPage: FC = () => {
       setBorehole(b);
       setEditingEnabled(b.locked !== null && b.lockedById === user.data.id);
     });
-  }, [id, user.data.id]);
+  }, [id, user.data.id, workflowStatus]);
 
   useEffect(() => {
     if (borehole?.updatedById) {