+++ src/client/src/pages/detail/form/borehole/boreholeGeneralSegment.tsx
@@ -2,9 +2,9 @@ import { useTranslation } from "react-i18next";
 import { Form, Segment, TextArea } from "semantic-ui-react";
 import DomainDropdown from "../../../../components/legacyComponents/domain/dropdown/domainDropdown.jsx";
 import { capitalizeFirstLetter } from "../../../../utils";
+import { BoreholeGeneralProps } from "./boreholePanelInterfaces.ts";
 
-const BoreholeGeneralSegment = props => {
-  const { borehole, updateChange, isEditable } = props;
+const BoreholeGeneralSegment = ({ legacyBorehole, updateChange, isEditable }: BoreholeGeneralProps) => {
   const { t } = useTranslation();
 
   return (
@@ -21,14 +21,14 @@ const BoreholeGeneralSegment = props => {
           <Form autoComplete="off" error>
             <Form.Group widths="equal">
               {/* drilling type in Borehole */}
-              <Form.Field error={borehole.data.borehole_type === null} required>
+              <Form.Field error={legacyBorehole.data.borehole_type === null} required>
                 <label>{t("borehole_type")}</label>
                 <DomainDropdown
-                  onSelected={selected => {
+                  onSelected={(selected: { id: number }) => {
                     updateChange("borehole_type", selected.id, false);
                   }}
                   schema="borehole_type"
-                  selected={borehole.data.borehole_type}
+                  selected={legacyBorehole.data.borehole_type}
                   readOnly={!isEditable}
                 />
               </Form.Field>
@@ -37,11 +37,11 @@ const BoreholeGeneralSegment = props => {
               <Form.Field required>
                 <label>{t("purpose")}</label>
                 <DomainDropdown
-                  onSelected={selected => {
+                  onSelected={(selected: { id: number }) => {
                     updateChange("extended.purpose", selected.id, false);
                   }}
                   schema="extended.purpose"
-                  selected={borehole.data.extended.purpose}
+                  selected={legacyBorehole.data.extended.purpose}
                   readOnly={!isEditable}
                 />
               </Form.Field>
@@ -50,11 +50,11 @@ const BoreholeGeneralSegment = props => {
               <Form.Field required>
                 <label>{capitalizeFirstLetter(t("boreholestatus"))}</label>
                 <DomainDropdown
-                  onSelected={selected => {
+                  onSelected={(selected: { id: number }) => {
                     updateChange("extended.status", selected.id, false);
                   }}
                   schema="extended.status"
-                  selected={borehole.data.extended.status}
+                  selected={legacyBorehole.data.extended.status}
                   readOnly={!isEditable}
                 />
               </Form.Field>
@@ -74,7 +74,7 @@ const BoreholeGeneralSegment = props => {
                   updateChange("custom.remarks", e.target.value);
                 }}
                 rows={14}
-                value={borehole.data.custom.remarks}
+                value={legacyBorehole.data.custom.remarks}
                 readOnly={!isEditable}
               />
             </Form.Field>