+++ src/client/src/pages/detail/labeling/labelingPanel.tsx
@@ -156,11 +156,8 @@ const LabelingPanel: FC<LabelingPanelProps> = ({ boreholeId }) => {
             }
           })
           .catch(error => {
-            if (!error?.message?.includes("AbortError")) {
+            if (!error?.toString().includes("AbortError")) {
               setExtractionState(ExtractionState.error);
-              // TODO: https://github.com/swisstopo/swissgeol-boreholes-suite/issues/1546
-              //  Check if error message is correct, resp. handle all error cases with different messages
-              //   Translate error message
               showAlert(t(error.message), "error");
             }
           })