+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - User permissions were not checked when detaching files from boreholes.
 - The basemap selector buttons were not displayed correctly on hover.
 - No scrollbar was displayed in the side drawer when many additional layers were added.
+- There was no appropriate error message when the drawn box did not contain extractable coordinates.
 
 ## v2.1.870 - 2024-09-27